Optimized Imports
Optimized imports and created a ReferenceFunctions and ReferenceConstants class to ensure consistency across commonly used functions and constants
