+++ operationportal/lib/Dashboards/BusDriver_ActiveDashboard.dart
@@ -1,27 +1,10 @@
-import 'dart:convert';
-
-import 'package:barcode_scan/barcode_scan.dart';
-import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Widget/AppBar.dart';
-import 'package:operationportal/ChildAddition/AddChild.dart';
-import 'package:operationportal/REST/Get_RetrieveBuses.dart';
-import 'package:operationportal/REST/Get_RetrieveChild.dart';
-import 'package:operationportal/REST/Get_RetrieveClasses.dart';
-import 'package:operationportal/REST/Get_RetrieveRoster.dart';
-import 'package:operationportal/REST/Get_RetrieveSuspendedRoster.dart';
-import 'package:operationportal/REST/Get_RetrieveUser.dart';
-import 'package:operationportal/REST/Post_ConfirmChildAttendance.dart';
-import 'package:operationportal/Structs/Child.dart';
-import 'package:operationportal/Structs/RosterChild.dart';
-import 'package:operationportal/Structs/Class.dart';
-import 'package:operationportal/Structs/Profile.dart';
-import 'package:operationportal/Structs/SuspendedChild.dart';
-import 'package:operationportal/Structs/User.dart';
 import 'package:flutter/material.dart';
+import 'package:operationportal/Structs/User.dart';
+import 'package:operationportal/Widget/BusDriverRoster.dart';
+import 'package:operationportal/Widget/ChildCheckIn.dart';
+import 'package:operationportal/Widget/Suspended.dart';
 
-import 'dart:async';
-
-import 'package:flutter/services.dart';
-import 'package:operationportal/Widget/SemiChildProfile.dart';
+import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Widget/AppBar.dart';
 
 import '../Structs/Storage.dart';
 
@@ -52,100 +35,8 @@ class BusDriver_ActiveDashboard_State extends State<BusDriver_ActiveDashboard_Pa
     Tab(text: 'Suspended'),
   ];
 
-  final List<int> colorCodes = <int>[600, 500];
-
   TabController _tabController;
-
-  void filterRosterResults(String query) {
-    if (children == null || childrenData == null)
-      return;
-
-    query = query.toUpperCase();
-
-    List<RosterChild> dummySearchList = List<RosterChild>();
-    dummySearchList.addAll(childrenData);
-    if(query.isNotEmpty) {
-      List<RosterChild> dummyListData = List<RosterChild>();
-      dummySearchList.forEach((item) {
-        if(item.firstName.toUpperCase().contains(query) || item.lastName.toUpperCase().contains(query)) {
-          dummyListData.add(item);
-        }
-      });
-      displayChildren.clear();
-      displayChildren.addAll(dummyListData);
-      setState(() {
-      });
-    } else {
-      displayChildren.clear();
-      setState(() {
-      });
-    }
-  }
-
-  void filterSuspendedResults(String query) {
-    if (suspended == null || suspendedData == null)
-      return;
-
-    query = query.toUpperCase();
-
-    List<RosterChild> dummySearchList = List<RosterChild>();
-    dummySearchList.addAll(suspendedData);
-    if(query.isNotEmpty) {
-      List<RosterChild> dummyListData = List<RosterChild>();
-      dummySearchList.forEach((item) {
-        if(item.firstName.toUpperCase().contains(query) || item.lastName.toUpperCase().contains(query)) {
-          dummyListData.add(item);
-        }
-      });
-      displaySuspended.clear();
-      displaySuspended.addAll(dummyListData);
-      setState(() {
-      });
-    } else {
-      displaySuspended.clear();
-      setState(() {
-      });
-    }
-  }
-
   Storage storage;
-  String token;
-
-  bool filterCheckedIn;
-
-  String barcode;
-
-  String busRouteValue;
-  String classIdValue;
-
-  final busRouteController = TextEditingController();
-  final classIdController = TextEditingController();
-
-  final rosterSearchController = TextEditingController();
-  final suspendedSearchController = TextEditingController();
-
-  List<String> busIds = new List<String>();
-  List<String> classIds = new List<String>();
-
-  List<RosterChild> displayChildren = new List<RosterChild>();
-  List<RosterChild> children = new List<RosterChild>();
-  List<RosterChild> childrenData = new List<RosterChild>();
-
-  List<RosterChild> displaySuspended = new List<RosterChild>();
-  List<RosterChild> suspended = new List<RosterChild>();
-  List<RosterChild> suspendedData = new List<RosterChild>();
-
-  DateTime parseBirthday (String birthday)
-  {
-    List<String> dateBreak = new List<String>();
-    dateBreak = birthday.split('/');
-    return DateTime(int.parse(dateBreak[2]), int.parse(dateBreak[0]), int.parse(dateBreak[1]));
-  }
-
-  int calculateBirthday(RosterChild child)
-  {
-    return DateTime.now().difference(parseBirthday(child.birthday.split(' ')[0])).inDays ~/ 365.25;
-  }
 
   Widget buildRefreshButton ()
   {
@@ -161,35 +52,12 @@ class BusDriver_ActiveDashboard_State extends State<BusDriver_ActiveDashboard_Pa
     );
   }
 
-  List<RosterChild> FilterChildren (List<RosterChild> children)
-  {
-    List<RosterChild> newList = new List<RosterChild>();
-    for (RosterChild c in children)
-    {
-      if (c.isCheckedIn)
-        newList.add(c);
-    }
-
-    return newList;
-  }
-
   @override
   void initState() {
-    super.initState();
     _tabController = TabController(vsync: this, length: myTabs.length);
-
-    busRouteValue = "3";
-    classIdValue = "Select Class";
-
-    filterCheckedIn = false;
-
-    busRouteController.text = busRouteValue;
-    classIdController.text = classIdValue;
-
     storage = new Storage();
-    storage.readToken().then((value) {
-      token = value;
-    });
+
+    super.initState();
   }
 
   @override
@@ -218,362 +86,14 @@ class BusDriver_ActiveDashboard_State extends State<BusDriver_ActiveDashboard_Pa
       controller: _tabController,
       children: myTabs.map((Tab tab) {
         if (tab.text == "Check-In")
-          {
-            return SingleChildScrollView(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: <Widget>[
-                  Container(
-                    child: IntrinsicHeight(
-                      child: Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: <Widget>
-                          [
-                            Flexible(
-                              child: Text("Recent Child", style: TextStyle(fontSize: 30, color: Colors.white),),
-                            ),
-                            Container(
-                              child: FlatButton(
-                                child: Text("Scan QR", style: TextStyle(color: Colors.black),),
-                                onPressed: () => scan(),
-                              ),
-                              decoration: new BoxDecoration(
-                                color: Colors.white,
-                                borderRadius: new BorderRadius.all(
-                                    new Radius.circular(20)
-                                ),
-                              ),
-                              margin: EdgeInsets.only(left: 20),
-                            ),
-                          ]
-                      ),
-                    ),
-                    decoration: new BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: new BorderRadius.all(
-                          new Radius.circular(20)
-                      ),
-                    ),
-                    padding: EdgeInsets.all(10),
-                    margin: EdgeInsets.all(10),
-                  ),
-                  FutureBuilder(
-                      future: RetrieveChild(token, barcode, context),
-                      builder: (BuildContext context, AsyncSnapshot<Child> snapshot) {
-                        switch (snapshot.connectionState) {
-                          case ConnectionState.none:
-                            return new Text('Issue Posting Data');
-                          case ConnectionState.waiting:
-                            return new Center(child: new CircularProgressIndicator());
-                          case ConnectionState.active:
-                            return new Text('');
-                          case ConnectionState.done:
-                            if (snapshot.hasError) {
-                              return Text("Press 'Scan QR' to begin!");
-                            } else {
-                              return Column(
-                                children: <Widget>[
-                                  Container(
-                                    child: IntrinsicHeight(
-                                      child: Row(
-                                          mainAxisAlignment: MainAxisAlignment.start,
-                                          crossAxisAlignment: CrossAxisAlignment.center,
-                                          children: <Widget>
-                                          [
-                                            Container(
-                                              child: CircleAvatar(
-                                                backgroundImage: (snapshot.data.picture != null) ? MemoryImage(base64.decode((snapshot.data.picture))) : null,
-                                              ),
-                                              height: 200,
-                                              width: 200,
-                                              padding: EdgeInsets.all(5),
-                                              margin: EdgeInsets.only(right: 10),
-                                            ),
-                                            Flexible(
-                                              child: Text(
-                                                snapshot.data.firstName + "\n" + snapshot.data.lastName,
-                                                textAlign: TextAlign.center,
-                                                style: TextStyle(fontSize: 28, color: Colors.white),
-                                              ),
-                                            ),
-                                          ]
-                                      ),
-                                    ),
-                                    decoration: new BoxDecoration(
-                                      color: Colors.blue,
-                                      borderRadius: new BorderRadius.all(
-                                          new Radius.circular(20)
-                                      ),
-                                    ),
-                                    padding: EdgeInsets.all(10),
-                                    margin: EdgeInsets.all(10),
-                                  ),
-                                  Container(
-                                      child: FlatButton(
-                                        child: Text("Confirm Attendance", style: TextStyle(fontSize: 20, color: Colors.white),),
-                                        onPressed: () {
-                                          ConfirmChildAttendance(token, snapshot.data, context);
-                                        },
-                                      ),
-                                      decoration: new BoxDecoration(
-                                        color: Colors.blue,
-                                        borderRadius: new BorderRadius.all(
-                                            new Radius.circular(20)
-                                        ),
-                                      ),
-                                      margin: EdgeInsets.only(top: 20)
-                                  ),
-                                ],
-                              );
-                            }
-                            break;
-                          default:
-                            return null;
-                        }
-                      }
-                  ),
-                ],
-              ),
-            );
-          }
+          return ChildCheckInPage(storage: storage,);
         else if(tab.text == "Roster")
-          return Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Container(
-                child: IntrinsicHeight(
-                  child: Row(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: <Widget>
-                      [
-                        Container(
-                          child: Text("Bus Route #$busRouteValue", textAlign: TextAlign.center,
-                            style: TextStyle(fontSize: 28, color: Colors.white),),
-                          decoration: new BoxDecoration(
-                            color: Colors.blue,
-                            borderRadius: new BorderRadius.all(
-                                new Radius.circular(20)
-                            ),
-                          ),
-                          padding: EdgeInsets.all(20),
-                          margin: EdgeInsets.only(right: 20),
-                        ),
-                      ]
-                  ),
-                ),
-                margin: EdgeInsets.all(10),
-              ),
-              Container(
-                child: CheckboxListTile(
-                  title: const Text('Filter Checked In'),
-                  value: filterCheckedIn,
-                  onChanged: (bool value) {
-                    setState(() { filterCheckedIn = !filterCheckedIn; });
-                  },
-                  secondary: const Icon(Icons.filter_tilt_shift),
-                ),
-              ),
-              Container(
-                child: IntrinsicHeight(
-                  child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: <Widget>
-                      [
-                        Flexible(
-                          child: TextField(
-                            onChanged: (value) {
-                              filterRosterResults(value);
-                            },
-                            controller: rosterSearchController,
-                            decoration: InputDecoration(
-                                labelText: "Search",
-                                prefixIcon: Icon(Icons.search),
-                                border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.all(Radius.circular(25.0)))
-                            ),
-                          ),
-                        ),
-                        Container(
-                            child: FlatButton(
-                                child: Text("Add Child"),
-                                onPressed: () => Navigator.push(context, MaterialPageRoute(builder: (context) => AddChildPage(title: 'Add Child')))
-                            )
-                        )
-                      ]
-                  ),
-                ),
-                margin: EdgeInsets.only(left: 10, bottom: 10),
-              ),
-              FutureBuilder(
-                  future: storage.readToken().then((value) {
-                    return RetrieveRoster(value, busRouteController.text, classIdController.text);
-                  }),
-                  builder: (BuildContext context, AsyncSnapshot<List<RosterChild>> snapshot) {
-                    switch (snapshot.connectionState) {
-                      case ConnectionState.none:
-                        return new Text('Issue Posting Data');
-                      case ConnectionState.waiting:
-                        return new Center(child: new CircularProgressIndicator());
-                      case ConnectionState.active:
-                        return new Text('');
-                      case ConnectionState.done:
-                        if (snapshot.hasError) {
-                          children = null;
-                          return Center(
-                            child: Text("Unable to Fetch Roster (May be an unassigned route!)"),
-                          );
-                        } else {
-                          childrenData = snapshot.data;
-                          children = displayChildren.length > 0 ? displayChildren : snapshot.data;
-                          filterCheckedIn ? children = FilterChildren(children) : children;
-                          return Expanded(
-                            child: new ListView.builder(
-                              itemCount: children.length,
-                              itemBuilder: (BuildContext context, int index) {
-                                return Container(
-                                  child: ListTile(
-                                    leading: Container(
-                                      child: CircleAvatar(
-                                        backgroundImage: (children[index].picture != null) ? MemoryImage(base64.decode((children[index].picture))) : null,
-                                      ),
-                                    ),
-                                    title: Text('${children[index].firstName} ' + '${children[index].lastName}',
-                                        style: TextStyle(color: Colors.white)),
-                                    subtitle: Text('${children[index].birthday != null && children[index].birthday.isNotEmpty ? 'Age: ' + '${calculateBirthday(children[index])}' : 'No Birthday Assigned'}', style: TextStyle(color: Colors.white)),
-                                    onTap: ()
-                                    {
-                                      Navigator.push(context, MaterialPageRoute(builder: (context) => SemiChildProfilePage(profile: widget.user.profile, child: children[index])));
-                                    },
-                                    dense: false,
-                                  ),
-                                  color: Colors.blue[colorCodes[index%2]],
-                                );
-                              },
-                            ),
-                          );
-                        }
-                        break;
-                      default:
-                        return null;
-                    }
-                  }
-              ),
-            ],
-          );
+          return BusRosterWidgetPage(storage: storage, user: widget.user,);
         else
-          return Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Container(
-                child: IntrinsicHeight(
-                  child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      crossAxisAlignment: CrossAxisAlignment.stretch,
-                      children: <Widget>
-                      [
-                        Flexible(
-                          child: TextField(
-                            onChanged: (value) {
-                              filterSuspendedResults(value);
-                            },
-                            controller: rosterSearchController,
-                            textAlign: TextAlign.left,
-                            decoration: InputDecoration(
-                                labelText: "Search",
-                                prefixIcon: Icon(Icons.search),
-                                border: OutlineInputBorder(
-                                    borderRadius: BorderRadius.all(Radius.circular(25.0)))),
-                            style: TextStyle(fontSize: 16),
-                          ),
-                        ),
-                      ]
-                  ),
-                ),
-                margin: EdgeInsets.all(10),
-              ),
-              FutureBuilder(
-                  future: storage.readToken().then((value) {
-                    return GetSuspendedChildren(value);
-                  }),
-                  builder: (BuildContext context, AsyncSnapshot<List<RosterChild>> snapshot) {
-                    switch (snapshot.connectionState) {
-                      case ConnectionState.none:
-                        return new Text('Issue Posting Data');
-                      case ConnectionState.waiting:
-                        return new Center(child: new CircularProgressIndicator());
-                      case ConnectionState.active:
-                        return new Text('');
-                      case ConnectionState.done:
-                        if (snapshot.hasError) {
-                          suspended = null;
-                          return Center(
-                            child: Text("Unable to Fetch Roster (May be an unassigned route!)"),
-                          );
-                        } else {
-                          suspendedData = snapshot.data;
-                          suspended = displaySuspended.length > 0 ? displaySuspended : snapshot.data;
-                          return Expanded(
-                            child: new ListView.builder(
-                              itemCount: suspended.length,
-                              itemBuilder: (BuildContext context, int index) {
-                                return Container(
-                                  child: ListTile(
-                                    leading: Container(
-                                      child: CircleAvatar(
-                                        backgroundImage: (suspended[index].picture != null) ? MemoryImage(base64.decode((suspended[index].picture))) : null,
-                                      ),
-                                    ),
-                                    title: Text('${suspended[index].firstName} ' + '${suspended[index].lastName}',
-                                        style: TextStyle(color: Colors.white)),
-                                    subtitle: Text('${suspended[index].birthday != null && suspended[index].birthday.isNotEmpty ? 'Age: ' + '${calculateBirthday(suspended[index])}' : 'No Birthday Assigned'}', style: TextStyle(color: Colors.white)),
-                                    onTap: ()
-                                    {
-                                      Navigator.push(context, MaterialPageRoute(builder: (context) => SemiChildProfilePage(profile: widget.user.profile, child: suspended[index])));
-                                    },
-                                    dense: false,
-                                  ),
-                                  color: Colors.blue[colorCodes[index%2]],
-                                );
-                              },
-                            ),
-                          );
-                        }
-                        break;
-                      default:
-                        return null;
-                    }
-                  }
-              ),
-            ],
-          );
+          return SuspendedWidgetPage(storage: storage, user: widget.user,);
       }).toList(),
     ),
     );
-  }
-
-  Future scan() async {
-    try {
-      String barcode = await BarcodeScanner.scan();
-      setState(() => this.barcode = barcode);
-    } on PlatformException catch (e) {
-      if (e.code == BarcodeScanner.CameraAccessDenied) {
-        setState(() {
-          this.barcode = 'The user did not grant the camera permission!';
-        });
-      } else {
-        setState(() => this.barcode = 'Unknown error: $e');
-      }
-    } on FormatException{
-      setState(() => this.barcode = 'null (User returned using the "back"-button before scanning anything. Result)');
-    } catch (e) {
-      setState(() => this.barcode = 'Unknown error: $e');
-    }
   }
 }
