+++ operationportal/lib/NoteAddition/AddNote.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:operationportal/REST/Post_CreateNote.dart';
-import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Structs/Storage.dart';
 import 'package:operationportal/Structs/Profile.dart';
 import 'package:operationportal/Structs/RosterChild.dart';
+
+import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Structs/Storage.dart';
 
 
 class AddNotePage extends StatefulWidget {
