+++ operationportal/lib/NoteAddition/NoteView.dart
@@ -1,8 +1,9 @@
+import 'package:flutter/material.dart';
 import 'package:operationportal/REST/Delete_Note.dart';
 import 'package:operationportal/REST/Post_EditNote.dart';
-import 'package:flutter/material.dart';
-import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Structs/Storage.dart';
 import 'package:operationportal/Structs/Note.dart';
+
+import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/Structs/Storage.dart';
 
 
 
@@ -22,7 +23,6 @@ class NoteViewState extends State<NoteViewPage>
   TextEditingController _authorController = new TextEditingController();
 
   bool noteControllerEditable;
-
 
   Storage storage;
   String originalNote;
