+++ operationportal/lib/REST/Delete_Note.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 
 Future<void> DeleteNote (String token, int id, BuildContext context)
