+++ operationportal/lib/REST/Get_RetrieveClassInfo.dart
@@ -1,10 +1,9 @@
 import 'dart:convert';
 
-import 'package:operationportal/Structs/Class.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
+import 'package:operationportal/Structs/Class.dart';
 
 Future<Class> RetrieveClassInfo (String token, String id, BuildContext context) async {
   var mUrl = "https://www.operation-portal.com/api/class-info?id=$id";
