+++ operationportal/lib/REST/Get_RetrieveNotes.dart
@@ -1,8 +1,7 @@
 import 'dart:convert';
 
-import 'package:operationportal/Structs/Note.dart';
 import 'package:http/http.dart' as http;
-import 'package:operationportal/Structs/RosterChild.dart';
+import 'package:operationportal/Structs/Note.dart';
 
 Future<List<Note>> RetrieveNotes (String token, int childId) async {
