+++ operationportal/lib/REST/Get_RetrieveVolunteerInfo.dart
@@ -6,7 +6,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 import 'package:operationportal/Structs/Volunteer.dart';
