+++ operationportal/lib/REST/Post_ConfirmVolunteerAttendance.dart
@@ -1,9 +1,7 @@
 import 'dart:convert';
 
-import 'package:operationportal/Structs/Profile.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 
 Future<void> ConfirmVolunteerAttendance (String token, int id, BuildContext context) async {
