+++ operationportal/lib/REST/Post_CreateInventory.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 
 Future<void> CreateInventory (String token, String name, int count, BuildContext context)
