+++ operationportal/lib/REST/Post_CreateNote.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 
 Future<void> CreateNote (String token, String author, int childId, String content, String priority, BuildContext context)
