+++ operationportal/lib/REST/Post_EditInventory.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
 
 Future<void> EditInventory (String token, String name, int count, int id, bool resolved, BuildContext context)
