+++ operationportal/lib/REST/Post_SuspendChild.dart
@@ -1,11 +1,9 @@
 import 'dart:convert';
 
-import 'package:operationportal/Structs/RosterChild.dart';
-import 'package:operationportal/Structs/SuspendedChild.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-
 import 'package:http/http.dart' as http;
+import 'package:operationportal/Structs/RosterChild.dart';
 
 Future<void> SuspendChild (String token, int id, String start, String end, BuildContext context)
 async {
