+++ operationportal/lib/References/ReferenceConstants.dart
@@ -1,0 +1,1 @@
+final List<int> colorCodes = <int>[600, 500];
