+++ operationportal/lib/Structs/RosterChild.dart
@@ -1,5 +1,3 @@
-import 'dart:typed_data';
-
 import 'Bus.dart';
 import 'Class.dart';
