+++ operationportal/lib/Structs/Volunteer.dart
@@ -1,6 +1,5 @@
 import 'Bus.dart';
 import 'Class.dart';
-import 'Profile.dart';
 
 class Volunteer {
   /*“volunteer”: {  “training”: [{“name”: “string”}], “referral”: “string”, “languages”: [“language”: “string”], “newsletter”: “bool”, “contactWhenShort”: “bool”, “phone”: “string”, “email”: “string”,“blueShirt”:bool, “nametag”:bool, “personalInterviewCompleted”:bool, “backgroundCheck”:bool, “yearStarted”:int, “canEditInventory”:bool, “picture”:byte[], “birthday”:DateTime}*/
