+++ operationportal/lib/Widget/Inventory.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
-import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/ItemAddition/AddItem.dart';
 import 'package:operationportal/REST/Get_RetrieveInventory.dart';
+import 'package:operationportal/References/ReferenceConstants.dart';
 import 'package:operationportal/Structs/Item.dart';
+
+import 'file:///C:/Users/gecco/Documents/GitHub/OperationPortal/operationportal/lib/ItemAddition/AddItem.dart';
 
 import '../ItemAddition/ItemView.dart';
 import '../Structs/Storage.dart';
@@ -18,8 +20,6 @@ class InventoryWidgetPage extends StatefulWidget
 
 class InventoryWidgetState extends State<InventoryWidgetPage>
 {
-  final List<int> colorCodes = <int>[600, 500];
-
   List<Item> displayItems = new List<Item>();
   List<Item> items = new List<Item>();
   List<Item> itemData = new List<Item>();
