+++ operationportal/lib/Widget/StaffWidgets/ClassProfile.dart
@@ -1,9 +1,6 @@
 
-import 'package:operationportal/Structs/Class.dart';
 import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
-
-import 'package:http/http.dart' as http;
+import 'package:operationportal/Structs/Class.dart';
 import 'package:operationportal/Structs/Storage.dart';
 
 class ClassProfilePage extends StatefulWidget {
