+++ operationportal/lib/Widget/SuspensionView.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import '../REST/Post_SuspendChild.dart';
-import '../Structs/Storage.dart';
 import '../Structs/RosterChild.dart';
+import '../Structs/Storage.dart';
 
 class SuspensionScreen extends StatefulWidget {
   SuspensionScreen({Key key, this.child}) : super(key: key);
