+++ operationportal/lib/Widget/TakePicture.dart
@@ -1,6 +1,4 @@
 // A screen that allows users to take a picture using a given camera.
-import 'dart:io';
-
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart';
