+++ operationportal/lib/Widget/TeacherRoster.dart
@@ -1,15 +1,15 @@
 import 'dart:convert';
 
-import 'package:operationportal/REST/Get_RetrieveClasses.dart';
-import 'package:operationportal/Structs/Bus.dart';
-import 'package:operationportal/Structs/Class.dart';
-import 'package:operationportal/Structs/Storage.dart';
-import 'package:operationportal/Structs/User.dart';
-import 'package:operationportal/ChildAddition/AddChild.dart';
-import 'package:operationportal/REST/Get_RetrieveRoster.dart';
-import 'package:operationportal/Structs/RosterChild.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:operationportal/ChildAddition/AddChild.dart';
+import 'package:operationportal/REST/Get_RetrieveRoster.dart';
+import 'package:operationportal/References/ReferenceConstants.dart';
+import 'package:operationportal/References/ReferenceFunctions.dart';
+import 'package:operationportal/Structs/Class.dart';
+import 'package:operationportal/Structs/RosterChild.dart';
+import 'package:operationportal/Structs/Storage.dart';
+import 'package:operationportal/Structs/User.dart';
 import 'package:operationportal/Widget/SemiChildProfile.dart';
 import 'package:operationportal/Widget/StaffWidgets/ClassProfile.dart';
 
@@ -26,8 +26,6 @@ class TeacherRosterWidgetPage extends StatefulWidget {
 
 class TeacherRosterWidgetState extends State<TeacherRosterWidgetPage>
 {
-  final List<int> colorCodes = <int>[600, 500];
-
   final classIdController = TextEditingController();
   int classIndex;
   List<String> classIds;
@@ -38,18 +36,6 @@ class TeacherRosterWidgetState extends State<TeacherRosterWidgetPage>
   List<RosterChild> displayChildren;
   List<RosterChild> children;
   List<RosterChild> childrenData;
-
-  DateTime parseBirthday (String birthday)
-  {
-    List<String> dateBreak = new List<String>();
-    dateBreak = birthday.split('/');
-    return DateTime(int.parse(dateBreak[2]), int.parse(dateBreak[0]), int.parse(dateBreak[1]));
-  }
-
-  int calculateBirthday(RosterChild child)
-  {
-    return DateTime.now().difference(parseBirthday(child.birthday.split(' ')[0])).inDays ~/ 365.25;
-  }
 
   void filterRosterResults(String query) {
     if (children == null || childrenData == null)
